fix(ui): change variable type for current time in userLedHandler to unsigned long

+++ examples/companion_radio/UITask.cpp
@@ -135,7 +135,7 @@ void UITask::userLedHandler() {
 #ifdef PIN_STATUS_LED
   static int state = 0;
   static int next_change = 0;
-  int cur_time = millis();
+  unsigned long cur_time = millis();
   if (cur_time > next_change) {
     if (state == 0) {
       state = 1; // led on, short = unread msg
